Moved createX methods into their own file. Updated createDraft function to always use timestamp if available
